[ScriptTracker] loadScriptFile handler update
Finished the first version of the loadScriptFile that should work.
Previous code was just initial thoughts and was not even complete.  Loading files based on object long ID was merged into the saveScriptFile handler which will import or export based on which side had changes.
Adjusted to take the file name as input and pull the long ID and stack name from the file.
Handler is not currently used.  Plan is to make use of it in the folder watch script to reduce time to process updates.