[Script Tracker] Add Prefs substack
Added a new substack to handle setting preferences
ScriptTracker preferences are stored in the same location as LiveCode preferences
Stack preferences are stored as a custom property set
One card configured to set overall settings, one for default paths, and one for the currently selected stack.
This will enable future preferences for things like opening scripts in an external editor.