[Script Tracker] Added sync on save option
Added a front script that will perform a sync before saving a stack if the option is selected.  Code is in the check box button for the option in the prefs stack.

As a part of that addition, added a fourth card to the preferences stack.  All of the code was moved from the individual cards to the stack level.  The controls/preferences updated for each card are now stored in custom properties for each card by control type.