[Script Tracker] Create new script files/Group editing mode fix
When external script editing is enabled, if an object without a script is edited, then a file is automatically generated so that it can be opened in an external editor.

Moved the external editor code into the behavior.  Revised to perform a sync before editing to ensure the current script is presented for editing.  Also updated the long ID/name that are generated to properly resolve objects that are part of a shared group.

Moved `ideSubscribe` to the `Automatic` button to remove one of the necessary checks (it will only be subscribed when the check is on).

Disable group editing mode when doing updates to ensure that the object long IDs are correct.  Calls are wrapped in a try due to errors when updating the stack that is currently running (but I don't use backgrounds, so it won't be a problem).