[Script Tracker] Rename collision option (script -> stack)
After review, "script" could be ambiguous.  Changing the option to "stack"
to be a little more clear.  The collision preference are now "stack", "file",
"skip", and "ask".  There still may be room to make it more clear, but for
backwards compatibility the original words/meanings will remain.

Performed some cleanup wrapping of long lines.

Added a check to be sure the stack is open before trying to save the stack
preferences.