[Script Tracker] Prevent SE from opening
Add a script to front to intercept calls to edit scripts.  Needed to include
two different handlers due to different paths that are taken when launching
the Script Editor.

* `editScript`
* `revIDEEditScriptOfObject`

Use the execution contexts to verify that when in browse mode the request
is from a contextual menu because otherwise it is usually going to be from
a request that is opening the SE for the debugger.

Updated log messages when a script is in use.  Visually have the "try later"
message appear under the "script changed" message.  The log file is still
in reverse chronological order otherwise.

Added an error message if unable to set the script of an object after the
delayed re-attempt.